sql文件以及日志操作记录

+++ common/src/main/java/model/dto/InsertUserDto.java
@@ -28,7 +28,7 @@ public class InsertUserDto {
 
     private String telephone;
 
-    private Integer mobile;
+    private Long mobile;
 
     private String email;
